Improve error handling on unknown error formats
If there is no `error_code` or `error_message`, we do the best we can by at least returning what was in the response content, pasted together as a character string.

This also explicitly uses `httr::content(response, "text")` for the bad response now, which is more robust than letting httr auto pick how to handle it.